# Conflicts:
#	libraries/applink/src/main/java/com/tokopedia/applink/ApplinkConst.java
#	libraries/applink/src/main/java/com/tokopedia/applink/internal/ApplinkConstInternalLogistic.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/marketplace/DeeplinkMapperLogistic.kt
#	libraries/applink/src/test/java/tokopedia/applink/deeplink/DeepLinkMapperCustomerAppTest.kt

+++ features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/filter/presentation/viewmodel/SomFilterViewModel.kt
@@ -75,13 +75,13 @@ class SomFilterViewModel @Inject constructor(dispatcher: CoroutineDispatchers,
         return getSomFilterUiModel(nameFilter)?.somFilterData.orEmpty()
     }
 
-    private fun findFirstSelectedFilterUiModelItemId(nameFilter: String): Int {
+    private fun findFirstSelectedFilterUiModelItemId(nameFilter: String): Long {
         return getSomFilterUiModelItems(nameFilter).find {
             it.isSelected
         }?.id.orZero()
     }
 
-    private fun findAllSelectedFilterUiModelItemsIds(nameFilter: String): MutableSet<Int> {
+    private fun findAllSelectedFilterUiModelItemsIds(nameFilter: String): MutableSet<Long> {
         return getSomFilterUiModelItems(nameFilter).filter {
             it.isSelected
         }.map {
